api/requirement.txt
#__author__ = "aki"

Django==3.0.3
django-cors-headers==3.2.1
django-encrypted-model-fields==0.5.8
django-filter==2.2.0
django-jsonfield==1.4.0
django-readonly-field==1.0.5
django-storages==1.9.1
djangorestframework==3.11.0
PyJWT==1.7.1
sentry-sdk==0.14.3
boto3==1.14.4
django-crontab==0.7.1
celery==4.4.5
redis==3.5.3
fsmpy==2.0.0
psycopg2==2.8.5
twilio==6.43.0
sendgrid==6.4.1
